5-dars/requirements.txt
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
requests==2.28.2
